move ClassDef to meta package
