+++ java/fury-core/src/main/java/org/apache/fury/builder/CodecUtils.java
@@ -24,10 +24,10 @@ import java.util.Collections;
 import org.apache.fury.Fury;
 import org.apache.fury.codegen.CodeGenerator;
 import org.apache.fury.codegen.CompileUnit;
+import org.apache.fury.meta.ClassDef;
 import org.apache.fury.resolver.ClassResolver;
 import org.apache.fury.resolver.FieldResolver;
 import org.apache.fury.serializer.Serializer;
-import org.apache.fury.type.ClassDef;
 import org.apache.fury.util.Preconditions;
 
 /** Codec util to create and load jit serializer class. */
