+++ java/fury-core/src/main/java/org/apache/fury/resolver/MetaContext.java
@@ -23,7 +23,7 @@ import java.util.ArrayList;
 import java.util.List;
 import org.apache.fury.collection.IdentityObjectIntMap;
 import org.apache.fury.memory.MemoryBuffer;
-import org.apache.fury.type.ClassDef;
+import org.apache.fury.meta.ClassDef;
 
 /**
  * Context for sharing class meta across multiple serialization. Class name, field name and field
