+++ java/fury-core/src/test/java/org/apache/fury/type/ClassDefTest.java
@@ -32,6 +32,7 @@ import java.util.TreeSet;
 import org.apache.fury.Fury;
 import org.apache.fury.FuryTestBase;
 import org.apache.fury.memory.MemoryBuffer;
+import org.apache.fury.meta.ClassDef;
 import org.apache.fury.util.ReflectionUtils;
 import org.testng.annotations.Test;
